- possible infinite loop fix

+++ dst_blacklist.c
@@ -464,7 +464,7 @@ inline static int dst_blacklist_clean_expired(unsigned int target,
 	int i;
 	
 	now=start_time=get_ticks_raw();
-	for(h=start; h!=(start+DST_BLST_HASH_SIZE); h++){
+	for(h=start; h!=(start+(unsigned short)DST_BLST_HASH_SIZE); h++){
 		i=h%DST_BLST_HASH_SIZE;
 		if (dst_blst_hash[i].first){
 			LOCK_BLST(i);
